% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_breaks}
\alias{base_breaks}
\title{Mimic Base R break}
\usage{
base_breaks(
  x,
  y,
  scale_x = T,
  scale_y = T,
  x_lab_fun = function(x) {     x },
  y_lab_fun = function(x) {     x },
  n_wrap = 10,
  outlier = 0.05,
  expand_x_conti = c(0.05, 0),
  expand_y_conti = c(0.05, 0),
  expand_x_disc = c(0, 0.6),
  expand_y_disc = c(0, 0.6)
)
}
\arguments{
\item{x}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions. default: "add_".
Note: You could set this the empty string, in which case the new functions would mask the name of the library function}

\item{y}{func_regex Regular expression to filter the list of ggplot functions to make pipe-enabled.  The default regex will capture all}
}
\description{
Mimic Base R break
}
