% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_wrap}
\alias{base_wrap}
\title{Warp list of plot as facets}
\usage{
base_wrap(plot.ls, nrow = "auto", ncol = "auto", labels = NULL, ...)
}
\arguments{
\item{p}{plot}

\item{facets}{a vector of column names used for facet}
}
\description{
Warp list of plot as facets
}
