% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smart_lab}
\alias{smart_lab}
\title{Mimic Base R break}
\usage{
smart_lab(lab)
}
\arguments{
\item{lab}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions. default: "add_".
Note: You could set this the empty string, in which case the new functions would mask the name of the library function
    stats and geoms and some other misc stuff.}
}
\description{
Mimic Base R break
}
