% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_tufte2}
\alias{theme_tufte2}
\title{Mimic Base R break}
\usage{
theme_tufte2(ticks = TRUE)
}
\arguments{
\item{x}{string with which to prefix names of \code{ggplot2} functions in order to name the pipe-enabled functions. default: "add_".
Note: You could set this the empty string, in which case the new functions would mask the name of the library function}

\item{y}{func_regex Regular expression to filter the list of ggplot functions to make pipe-enabled.  The default regex will capture all}
}
\description{
Mimic Base R break
}
